Sprint 3 increment 11: DB storage
* Implement the PostgreSQL storage.
* Implement shared unit-test cases for all type of storages.
* Put the storage-specific tests into separate files.
